Do not force dependants to depend on bazel_gazelle
By loading the "gazelle" rule from @bazel_gazelle in the top-level build
file, all users of targets in that file have to depend on bazel_gazelle.
This is particular problematic for rules_go, which depends on
//:validators_golang but should not depend on bazel_gazelle.

This is resolved by moving the gazelle rule definition into a separate
Bazel package.